backend/requirements.txt
fastapi
uvicorn
web3
python-dotenv
eth-account
cryptography
